userID: break out EID configuration to individual submodules (#10229)
* move eids config to submodules

* update EID logic

* Fix tests
