+++ modules/uid2IdSystem.js
@@ -83,6 +83,20 @@ export const uid2IdSubmodule = {
     _logInfo(`UID2 getId returned`, result);
     return result;
   },
+  eids: {
+    'uid2': {
+      source: 'uidapi.com',
+      atype: 3,
+      getValue: function(data) {
+        return data.id;
+      },
+      getUidExt: function(data) {
+        if (data.ext) {
+          return data.ext;
+        }
+      }
+    },
+  },
 };
 
 function decodeImpl(value) {
