+++ test/spec/modules/rubiconBidAdapter_spec.js
@@ -270,7 +270,81 @@ describe('the rubicon adapter', function () {
       }],
       criteoId: '1111',
     };
-    bid.userIdAsEids = createEidsArray(bid.userId);
+    bid.userIdAsEids = [
+      {
+        'source': 'liveintent.com',
+        'uids': [
+          {
+            'id': '0000-1111-2222-3333',
+            'atype': 3
+          }
+        ],
+        'ext': {
+          'segments': [
+            'segA',
+            'segB'
+          ]
+        }
+      },
+      {
+        'source': 'liveramp.com',
+        'uids': [
+          {
+            'id': '1111-2222-3333-4444',
+            'atype': 3
+          }
+        ]
+      },
+      {
+        'source': 'adserver.org',
+        'uids': [
+          {
+            'id': '3000',
+            'atype': 1,
+            'ext': {
+              'rtiPartner': 'TDID'
+            }
+          }
+        ]
+      },
+      {
+        'source': 'pubcid.org',
+        'uids': [
+          {
+            'id': '4000',
+            'atype': 1
+          }
+        ]
+      },
+      {
+        'source': 'example.com',
+        'uids': [
+          {
+            'id': '333333',
+            'ext': {
+              'stype': 'ppuid'
+            }
+          }
+        ]
+      },
+      {
+        'source': 'id-partner.com',
+        'uids': [
+          {
+            'id': '4444444'
+          }
+        ]
+      },
+      {
+        'source': 'criteo.com',
+        'uids': [
+          {
+            'id': '1111',
+            'atype': 1
+          }
+        ]
+      }
+    ];
     return bidderRequest;
   }
 
@@ -1224,7 +1298,20 @@ describe('the rubicon adapter', function () {
             clonedBid.userId = {
               tdid: 'abcd-efgh-ijkl-mnop-1234'
             };
-            clonedBid.userIdAsEids = createEidsArray(clonedBid.userId);
+            clonedBid.userIdAsEids = [
+              {
+                'source': 'adserver.org',
+                'uids': [
+                  {
+                    'id': 'abcd-efgh-ijkl-mnop-1234',
+                    'atype': 1,
+                    'ext': {
+                      'rtiPartner': 'TDID'
+                    }
+                  }
+                ]
+              }
+            ];
             let [request] = spec.buildRequests([clonedBid], bidderRequest);
             let data = parseQuery(request.data);
 
@@ -1241,7 +1328,23 @@ describe('the rubicon adapter', function () {
                   segments: ['segA', 'segB']
                 }
               };
-              clonedBid.userIdAsEids = createEidsArray(clonedBid.userId);
+              clonedBid.userIdAsEids = [
+                {
+                  'source': 'liveintent.com',
+                  'uids': [
+                    {
+                      'id': '0000-1111-2222-3333',
+                      'atype': 3
+                    }
+                  ],
+                  'ext': {
+                    'segments': [
+                      'segA',
+                      'segB'
+                    ]
+                  }
+                }
+              ];
               let [request] = spec.buildRequests([clonedBid], bidderRequest);
               let data = parseQuery(request.data);
 
@@ -1258,7 +1361,23 @@ describe('the rubicon adapter', function () {
                   segments: ['segD', 'segE']
                 }
               };
-              clonedBid.userIdAsEids = createEidsArray(clonedBid.userId);
+              clonedBid.userIdAsEids = [
+                {
+                  'source': 'liveintent.com',
+                  'uids': [
+                    {
+                      'id': '1111-2222-3333-4444',
+                      'atype': 3
+                    }
+                  ],
+                  'ext': {
+                    'segments': [
+                      'segD',
+                      'segE'
+                    ]
+                  }
+                }
+              ]
               let [request] = spec.buildRequests([clonedBid], bidderRequest);
               const unescapedData = unescape(request.data);
 
@@ -1273,7 +1392,17 @@ describe('the rubicon adapter', function () {
               clonedBid.userId = {
                 idl_env: '1111-2222-3333-4444'
               };
-              clonedBid.userIdAsEids = createEidsArray(clonedBid.userId);
+              clonedBid.userIdAsEids = [
+                {
+                  'source': 'liveramp.com',
+                  'uids': [
+                    {
+                      'id': '1111-2222-3333-4444',
+                      'atype': 3
+                    }
+                  ]
+                }
+              ]
               let [request] = spec.buildRequests([clonedBid], bidderRequest);
               let data = parseQuery(request.data);
 
@@ -1287,7 +1416,17 @@ describe('the rubicon adapter', function () {
               clonedBid.userId = {
                 pubcid: '1111'
               };
-              clonedBid.userIdAsEids = createEidsArray(clonedBid.userId);
+              clonedBid.userIdAsEids = [
+                {
+                  'source': 'pubcid.org',
+                  'uids': [
+                    {
+                      'id': '1111',
+                      'atype': 1
+                    }
+                  ]
+                }
+              ]
               let [request] = spec.buildRequests([clonedBid], bidderRequest);
               let data = parseQuery(request.data);
 
@@ -1301,7 +1440,17 @@ describe('the rubicon adapter', function () {
               clonedBid.userId = {
                 criteoId: '1111'
               };
-              clonedBid.userIdAsEids = createEidsArray(clonedBid.userId);
+              clonedBid.userIdAsEids = [
+                {
+                  'source': 'criteo.com',
+                  'uids': [
+                    {
+                      'id': '1111',
+                      'atype': 1
+                    }
+                  ]
+                }
+              ]
               let [request] = spec.buildRequests([clonedBid], bidderRequest);
               let data = parseQuery(request.data);
 
@@ -1328,7 +1477,27 @@ describe('the rubicon adapter', function () {
                   }]
                 }]
               };
-              clonedBid.userIdAsEids = createEidsArray(clonedBid.userId);
+              clonedBid.userIdAsEids = [
+                {
+                  'source': 'example.com',
+                  'uids': [
+                    {
+                      'id': '11111',
+                      'ext': {
+                        'stype': 'ppuid'
+                      }
+                    }
+                  ]
+                },
+                {
+                  'source': 'id-partner.com',
+                  'uids': [
+                    {
+                      'id': '222222'
+                    }
+                  ]
+                }
+              ];
               let [request] = spec.buildRequests([clonedBid], bidderRequest);
               let data = parseQuery(request.data);
 
@@ -1347,7 +1516,20 @@ describe('the rubicon adapter', function () {
                   }
                 }
               };
-              clonedBid.userIdAsEids = createEidsArray(clonedBid.userId);
+              clonedBid.userIdAsEids = [
+                {
+                  'source': 'id5-sync.com',
+                  'uids': [
+                    {
+                      'id': '11111',
+                      'atype': 1,
+                      'ext': {
+                        'linkType': '22222'
+                      }
+                    }
+                  ]
+                }
+              ];
               let [request] = spec.buildRequests([clonedBid], bidderRequest);
               let data = parseQuery(request.data);
