+++ test/spec/modules/smaatoBidAdapter_spec.js
@@ -1095,17 +1095,15 @@ describe('smaatoBidAdapterTest', () => {
             criteoId: '123456',
             tdid: '89145'
           },
-          userIdAsEids: createEidsArray({
-            criteoId: '123456',
-            tdid: '89145'
-          })
+          userIdAsEids: [
+            {id: 1}, {id: 2}
+          ]
         };
 
         const reqs = spec.buildRequests([userIdBidRequest], defaultBidderRequest);
 
         const req = extractPayloadOfFirstAndOnlyRequest(reqs);
-        expect(req.user.ext.eids).to.exist;
-        expect(req.user.ext.eids).to.have.length(2);
+        expect(req.user.ext.eids).to.eql(userIdBidRequest.userIdAsEids);
       });
     });
